samples/requirements.txt
# Command to Run:  pip3 install -r  requirements.txt 
# Core dependencies
# Flask==2.3.3
psycopg2-binary==2.9.7
# Flask-SQLAlchemy==3.0.5
# Flask-Migrate==4.0.4
# Flask-Cors==4.0.0
# Flask-JWT-Extended==4.5.2
# Flask-Mail==0.9.1
# Werkzeug==2.3.7
# python-dotenv==1.0.0
# email-validator==2.0.0.post2

# For production
# gunicorn==21.2.0

# # Development tools
# pytest==7.4.0
# black==23.7.0
# flake8==6.1.0
